upload day 1 (easy)
create a program that will ask the users name, age, and reddit username.
have it tell them the information back, in the format:
your name is (blank), you are (blank) years old, and your username is
(blank)
for extra credit, have the program log this information in a file to be
accessed later.