upload day 1 (hard)
we all know the classic "guessing game" with higher or lower prompts.
lets do a role reversal; you create a program that will guess numbers
between 1-100, and respond appropriately based on whether users say that
the number is too high or too low. Try to make a program that can guess
your number based on user input and great code!